moved get crs from lat long to ee_download
